A start-from-scratch redo of the core worker loop, in the form of BlockTransactions.
Unlike ResumableTransactions, this accepts a full transaction over GRPC
as a block.  This is a much nicer fit to the standard transactions model
of running a block of logic in a lambda, and fixes a lot of issues seen
with ResumableTransactions.

ResumableTransactions has been left for now, as not all tests have been
ported to BlockTransactions, but is likely to be removed soon.